google login and password recoveri

+++ src/main/java/com/wrap/it/dto/item/cart/CreateCartItemRequestDto.java
@@ -10,5 +10,5 @@ public class CreateCartItemRequestDto {
     @NotNull
     private Long itemId;
     @Positive(message = "Quantity must be at least 1")
-    private int quantity;
+    private short quantity;
 }
